requirements.txt
bcrypt==4.2.0
SQLAlchemy==2.0.32
asyncpg==0.29.0
uvicorn==0.30.5
fastapi==0.112.0
fastapi-jwt-auth==0.5.0
PyJWT~=1.7.1
pydantic~=1.10.8
pytest==8.3.2
